# Make #dig iterative