Fixes #2300. Related to #2299.

Authored with @jantnovi.