@eregon 2021-03-16
I think we really need benchmarks to drive this.
The approach with `each_with_index` and a block might actually be slower than what we have currently.