@chrisseaton 2021-03-17
Hmmm it's complicated. But you can see the performance cliff as. I think there's some things we can do to add overhead we added that impacts the interpreter.

Reference

```
    ruby 2.7.2p137 (2020-10-01 revision 5445e04352) [x86_64-darwin19]
    Warming up --------------------------------------
                dig-01     1.480M i/100ms
                dig-02     1.061M i/100ms
                dig-03   984.798k i/100ms
                dig-04   861.577k i/100ms
                dig-05   934.118k i/100ms
                dig-06   764.349k i/100ms
                dig-07   713.700k i/100ms
                dig-08   690.318k i/100ms
                dig-09   612.616k i/100ms
                dig-10   605.357k i/100ms
    Calculating -------------------------------------
                dig-01     12.726M (± 7.8%) i/s -     63.624M in   5.030961s
                dig-02     12.318M (± 3.5%) i/s -     61.514M in   5.000037s
                dig-03     12.731M (± 3.6%) i/s -     64.012M in   5.034868s
                dig-04     11.973M (± 6.1%) i/s -     60.310M in   5.059352s
                dig-05     11.391M (± 4.7%) i/s -     56.981M in   5.014500s
                dig-06     10.768M (± 4.6%) i/s -     54.269M in   5.052218s
                dig-07     10.177M (± 6.2%) i/s -     50.673M in   5.001752s
                dig-08      9.295M (±10.0%) i/s -     46.251M in   5.052569s
                dig-09      9.043M (± 5.3%) i/s -     45.334M in   5.031412s
                dig-10      8.596M (± 6.0%) i/s -     42.980M in   5.021170s

    Comparison:
                dig-03: 12730818.7 i/s
                dig-01: 12726054.6 i/s - same-ish: difference falls within error
                dig-02: 12317872.6 i/s - same-ish: difference falls within error
                dig-04: 11972985.5 i/s - same-ish: difference falls within error
                dig-05: 11391095.7 i/s - 1.12x  (± 0.00) slower
                dig-06: 10768281.0 i/s - 1.18x  (± 0.00) slower
                dig-07: 10177428.7 i/s - 1.25x  (± 0.00) slower
                dig-08:  9294513.6 i/s - 1.37x  (± 0.00) slower
                dig-09:  9042881.7 i/s - 1.41x  (± 0.00) slower
                dig-10:  8596371.1 i/s - 1.48x  (± 0.00) slower
```

Before:

```
    truffleruby 21.1.0-dev-d3ff03e9, like ruby 2.7.2, Interpreted JVM [x86_64-darwin]
    Warming up --------------------------------------
                dig-01    99.808k i/100ms
                dig-02    56.122k i/100ms
                dig-03    37.177k i/100ms
                dig-04    26.799k i/100ms
                dig-05    21.294k i/100ms
                dig-06    18.375k i/100ms
                dig-07    15.861k i/100ms
                dig-08    13.843k i/100ms
                dig-09    11.991k i/100ms
                dig-10    10.782k i/100ms
    Calculating -------------------------------------
                dig-01      1.129M (± 4.0%) i/s -      5.689M in   5.050132s
                dig-02    522.579k (± 9.6%) i/s -      2.582M in   5.003827s
                dig-03    367.592k (± 2.2%) i/s -      1.859M in   5.059331s
                dig-04    270.424k (± 3.9%) i/s -      1.367M in   5.062899s
                dig-05    214.627k (± 3.8%) i/s -      1.086M in   5.067702s
                dig-06    175.038k (± 6.1%) i/s -    882.000k in   5.061318s
                dig-07    150.371k (± 4.5%) i/s -    761.328k in   5.074351s
                dig-08    133.352k (± 3.3%) i/s -    678.307k in   5.092764s
                dig-09    116.988k (± 6.6%) i/s -    587.559k in   5.048467s
                dig-10    107.362k (± 6.0%) i/s -    539.100k in   5.045608s

    Comparison:
                dig-01:  1128520.0 i/s
                dig-02:   522579.4 i/s - 2.16x  (± 0.00) slower
                dig-03:   367592.4 i/s - 3.07x  (± 0.00) slower
                dig-04:   270424.2 i/s - 4.17x  (± 0.00) slower
                dig-05:   214626.8 i/s - 5.26x  (± 0.00) slower
                dig-06:   175037.7 i/s - 6.45x  (± 0.00) slower
                dig-07:   150371.3 i/s - 7.50x  (± 0.00) slower
                dig-08:   133351.9 i/s - 8.46x  (± 0.00) slower
                dig-09:   116987.9 i/s - 9.65x  (± 0.00) slower
                dig-10:   107361.6 i/s - 10.51x  (± 0.00) slower

    truffleruby 21.1.0-dev-d3ff03e9, like ruby 2.7.2, GraalVM CE JVM [x86_64-darwin]
    Warming up --------------------------------------
                dig-01    40.362M i/100ms
                dig-02    21.260M i/100ms
                dig-03    13.711M i/100ms
                dig-04     3.566M i/100ms
                dig-05     1.077M i/100ms
                dig-06   639.318k i/100ms
                dig-07   443.567k i/100ms
                dig-08   344.086k i/100ms
                dig-09   266.934k i/100ms
                dig-10   222.807k i/100ms
    Calculating -------------------------------------
                dig-01    386.621M (±11.3%) i/s -      1.897B in   5.033446s
                dig-02    210.387M (± 2.2%) i/s -      1.063B in   5.055128s
                dig-03    136.705M (± 2.7%) i/s -    685.560M in   5.018844s
                dig-04     37.247M (± 4.5%) i/s -    188.997M in   5.085941s
                dig-05     10.838M (± 4.0%) i/s -     54.948M in   5.078872s
                dig-06      6.234M (± 7.9%) i/s -     31.327M in   5.069907s
                dig-07      4.310M (± 4.6%) i/s -     21.735M in   5.054481s
                dig-08      3.310M (± 3.1%) i/s -     16.860M in   5.099078s
                dig-09      2.516M (±10.9%) i/s -     12.546M in   5.058044s
                dig-10      2.152M (± 7.8%) i/s -     10.695M in   5.006918s

    Comparison:
                dig-01: 386620520.5 i/s
                dig-02: 210386963.5 i/s - 1.84x  (± 0.00) slower
                dig-03: 136704634.3 i/s - 2.83x  (± 0.00) slower
                dig-04: 37246992.7 i/s - 10.38x  (± 0.00) slower
                dig-05: 10838474.8 i/s - 35.67x  (± 0.00) slower
                dig-06:  6234175.4 i/s - 62.02x  (± 0.00) slower
                dig-07:  4310394.6 i/s - 89.69x  (± 0.00) slower
                dig-08:  3309889.0 i/s - 116.81x  (± 0.00) slower
                dig-09:  2515960.2 i/s - 153.67x  (± 0.00) slower
                dig-10:  2151574.2 i/s - 179.69x  (± 0.00) slower
```

After:

```
    truffleruby 21.1.0-dev-d3ff03e9, like ruby 2.7.2, Interpreted JVM [x86_64-darwin]
    Warming up --------------------------------------
                dig-01    95.928k i/100ms
                dig-02    22.407k i/100ms
                dig-03    14.126k i/100ms
                dig-04     9.486k i/100ms
                dig-05     9.198k i/100ms
                dig-06     7.639k i/100ms
                dig-07     6.585k i/100ms
                dig-08     5.850k i/100ms
                dig-09     5.173k i/100ms
                dig-10     4.614k i/100ms
    Calculating -------------------------------------
                dig-01      1.086M (± 7.8%) i/s -      5.468M in   5.071921s
                dig-02    209.206k (± 8.0%) i/s -      1.053M in   5.076536s
                dig-03    144.173k (± 4.8%) i/s -    720.426k in   5.009842s
                dig-04    107.728k (± 6.8%) i/s -    540.702k in   5.045705s
                dig-05     89.924k (± 4.1%) i/s -    450.702k in   5.021477s
                dig-06     74.129k (± 5.2%) i/s -    374.311k in   5.064742s
                dig-07     62.772k (± 7.1%) i/s -    316.080k in   5.064829s
                dig-08     55.363k (± 6.0%) i/s -    280.800k in   5.092158s
                dig-09     49.587k (± 5.4%) i/s -    248.304k in   5.023728s
                dig-10     44.195k (± 6.3%) i/s -    221.472k in   5.033675s

    Comparison:
                dig-01:  1085661.5 i/s
                dig-02:   209205.8 i/s - 5.19x  (± 0.00) slower
                dig-03:   144173.1 i/s - 7.53x  (± 0.00) slower
                dig-04:   107728.2 i/s - 10.08x  (± 0.00) slower
                dig-05:    89924.4 i/s - 12.07x  (± 0.00) slower
                dig-06:    74128.9 i/s - 14.65x  (± 0.00) slower
                dig-07:    62772.3 i/s - 17.30x  (± 0.00) slower
                dig-08:    55363.3 i/s - 19.61x  (± 0.00) slower
                dig-09:    49587.5 i/s - 21.89x  (± 0.00) slower
                dig-10:    44195.3 i/s - 24.57x  (± 0.00) slower

    truffleruby 21.1.0-dev-d3ff03e9, like ruby 2.7.2, GraalVM CE JVM [x86_64-darwin]
    Warming up --------------------------------------
                dig-01    40.281M i/100ms
                dig-02    22.851M i/100ms
                dig-03    12.737M i/100ms
                dig-04   954.148k i/100ms
                dig-05     7.454M i/100ms
                dig-06     6.547M i/100ms
                dig-07     5.544M i/100ms
                dig-08     4.666M i/100ms
                dig-09   943.786k i/100ms
                dig-10     3.424M i/100ms
    Calculating -------------------------------------
                dig-01    386.739M (±11.5%) i/s -      1.893B in   5.027630s
                dig-02    234.426M (± 2.7%) i/s -      1.188B in   5.072749s
                dig-03    155.197M (± 1.2%) i/s -    776.966M in   5.007096s
                dig-04    108.056M (± 5.8%) i/s -    538.139M in   5.000582s
                dig-05     83.157M (± 2.2%) i/s -    417.416M in   5.022234s
                dig-06     64.157M (± 2.0%) i/s -    320.823M in   5.002716s
                dig-07     54.648M (± 2.8%) i/s -    277.184M in   5.076330s
                dig-08     45.555M (± 2.9%) i/s -    228.653M in   5.024044s
                dig-09     38.667M (± 6.0%) i/s -    192.532M in   4.999830s
                dig-10     33.911M (± 4.8%) i/s -    171.211M in   5.061707s

    Comparison:
                dig-01: 386738844.3 i/s
                dig-02: 234426182.4 i/s - 1.65x  (± 0.00) slower
                dig-03: 155197411.3 i/s - 2.49x  (± 0.00) slower
                dig-04: 108055746.9 i/s - 3.58x  (± 0.00) slower
                dig-05: 83157089.9 i/s - 4.65x  (± 0.00) slower
                dig-06: 64157470.4 i/s - 6.03x  (± 0.00) slower
                dig-07: 54648123.6 i/s - 7.08x  (± 0.00) slower
                dig-08: 45554694.5 i/s - 8.49x  (± 0.00) slower
                dig-09: 38666689.2 i/s - 10.00x  (± 0.00) slower
                dig-10: 33911419.4 i/s - 11.40x  (± 0.00) slower
```